Stripped everything away except for the UnknownMatrix code. (#7)
Most of the parsing work will be done by beachmat, as it is easier to do this
in R than in C++. Similarly, the main thread execution code is now handled by
manticore, meaning that the (renamed) tatami_r library is only responsible for
the tatami::UnknownMatrix definition. This greatly simplifies the code.