Enable oracle-aware extraction from the UnknownMatrix. (#8)
This uses the oracle to extract only the elements of interest during the
extract_array() call. We don't do any granular chunk management because it's
too costly to split the extracted array into its chunks, especially in the
consecutive case. Rather, at each cycle of predictions, we just do a single
call into R and hold onto the resulting matrix as an invisible unit. This
reduces the number of calls into R, at the cost of re-loading the same data if
the current cycle of predictions shares many elements with the previous cycle. 

If chunkdim() != NULL, we "round up" the cache usage of each predicted element
to the chunk size if the predicted element belongs in a previously unseen chunk
in this prediction cycle. This aims to avoid partially reading a chunk in one
prediction cycle, and then re-reading the rest of the chunk in another cycle,
which would be wasteful, e.g., for file-backed matrices.