Respect the index order of the returned SparseArraySeed.
If the returned object is already CSC or CSR, we just use that directly rather
than performing another sort. However, if the returned object contains unsorted
indices, and we need to sort anyway, we convert it to the preferred layout for
the iteration dimension, e.g., CSR for row-wise iteration.