Update classes to match the new tatami interface, improve caching. (#10)
The Extractor interface updates allow users to directly pass an oracle during
construction, which eliminates the messy work required to choose between the
with- and without-oracle behavior after construction. We also use the new
SparsifiedWrapper classes for efficiently dense output from a sparse matrix.

We improve chunk caching by accepting arbitrary chunkGrids rather than being
limited to regular chunk intervals. We simplify cache management by not
considering R-allocated memory in the cache size limit, which allows us to copy
over the data into our own slab cache managers from tatami_chunked. This
allows us to actually use the LRU and oracular caches to their full extent.

We also remove the COO matrix parser, on the understanding that this is superceded
by the SVT matrices in the latest versions of DelayedArray. (Nonetheless, we add
an R-side fallback to convert COO to SVT just in case.)

Finally, we greatly enhanced the scale of the tests to cover the new code.